Repo restructure
- Restructured repository.
- Small README changes
