--- 3.single_modular_analysis/README.md
@@ -1,13 +1,0 @@
-# 3. Single-Modular Analysis
-This subsection of the repository contains the single-modular main analysis. For an extensive elaboration on the approach, please consult the research manuscript [manuscript.pdf](https://github.com/BartJanBoverhof/Masterthesis/tree/main/1.latex_manuscript).
-
----
-
-![Status](https://img.shields.io/static/v1?label=Code+Status&message=Unfinished+and+Unexcecutable&color=red) 
-
----
-
-This subsection of the repository contains the following objects: 
-* `eeg_net.py`: Python script that defines the EEG-modality ConvNet structure.
-* `ppg_net.py`: Python script that defines the PPG-modality ConvNet structure.
-* `gsr_net.py`: Python script that defines the GSR-modality ConvNet structure.
