--- 4.multi_modular_analysis/README.md
@@ -1,11 +1,0 @@
-# 4. Multi-Modular Analysis
-This subsection of the repository contains the multi-modular main analysis. For an extensive elaboration on the approach, please consult the research manuscript [manuscript.pdf](https://github.com/BartJanBoverhof/Masterthesis/tree/main/1.latex_manuscript).
-
----
-
-![Status](https://img.shields.io/static/v1?label=Code+Status&message=Unfinished+and+Unexcecutable&color=red) 
-
----
-
-This subsection of the repository contains the following objects: 
-* `main_network.py`: Python script that defines the EEG-modality ConvNet structure.
