+++ manuscript/figures/experimental_setup.png
