+++ manuscript/figures/multi_model_architecture.png
