+++ manuscript/figures/single_model_architecture.png
