+++ manuscript/manuscript.synctex.gz
