+++ manuscript/manuscript.tex
