+++ pipeline/main.py
@@ -8,10 +8,17 @@
 import pyxdf
 import pandas as pd
 import pickle
-import sys
 import os
 
-from transformer.py import transformer
+#Setting WD to "Masterthesis/5.pipeline" folder. 
+os.chdir("pipeline")
+
+try:
+    from transformer import transformer
+except ModuleNotFoundError:
+    wd = os.getcwd()
+    print("Please make sure that working directory is set as '../Masterthesis/pipeline'")
+    print("Current working directory is:", wd)
 
 #### Preliminaries ####
 #Defining required events and streams
@@ -27,21 +34,18 @@ window_exact = True
 
 
 
-
-
 if __name__ == '__main__': 
     #First segment
-    seg1 = main("5.pipeline/data/bc10/bci10 operations.xdf") #Cut first segment
-    with open('5.pipeline/prepared_data/bci10_1.pickle', 'wb') as handle: #Save as .pickle
+    seg1 = transformer("data/bc10/bci10 operations.xdf") #Cut first segment
+    with open('prepared_data/10_1.pickle', 'wb') as handle: #Save as .pickle
         pickle.dump(seg1, handle, protocol=pickle.HIGHEST_PROTOCOL)
 
     #Second segment
-    seg2 = main("5.pipeline/data/bc10/bci 10 engineering+.xdf")
-    with open('5.pipeline/prepared_data/bci10_2.pickle', 'wb') as handle: #Save as .pickle
+    seg2 = transformer("data/bc10/bci 10 engineering+.xdf")
+    with open('prepared_data/10_2.pickle', 'wb') as handle: #Save as .pickle
         pickle.dump(seg2, handle, protocol=pickle.HIGHEST_PROTOCOL)
 
     #Third segment
-    seg3 = main("5.pipeline/data/bc10/bci10 tactical.xdf")
-    with open('5.pipeline/prepared_data/bci10_3.pickle', 'wb') as handle: #Save as .pickle
+    seg3 = transformer("data/bc10/bci10 tactical.xdf")
+    with open('prepared_data/10_3.pickle', 'wb') as handle: #Save as .pickle
         pickle.dump(seg3, handle, protocol=pickle.HIGHEST_PROTOCOL)
-
