+++ pytorch/README.md
@@ -1,0 +1,18 @@
+# PyTorch
+This subsection of the repository contains all PyTorch related code. This includes network architectures utilized in the main analysis, as well as scripts that are utilized for training.
+
+For an extensive elaboration on the approach, please consult the research manuscript [manuscript.pdf](https://github.com/BartJanBoverhof/Masterthesis/tree/main/1.latex_manuscript).
+
+---
+
+![Status](https://img.shields.io/static/v1?label=Code+Status&message=Unfinished+and+Unexcecutable&color=red) 
+
+---
+
+This subsection of the repository contains the following objects: 
+* `networks.py`: Script defining the following network definitions:
+    *  `EegNet`: Single modular EEG network architecture.
+    *  `PpgNet`: Single modular PPG network architecture.
+    *  `GsrNet`: Single modular GSR network architecture.
+    *  `MultiNet`: Multi modular network architecture.
+* `train.py`: Python script with which networks are trained.
